requirements.txt
# pip install aws-cdk
aws-cdk.core==1.85.0
aws-cdk.aws-dynamodb==1.85.0
aws-cdk.aws-ec2==1.85.0
aws-cdk.aws-iam==1.85.0
aws-cdk.aws-lambda==1.85.0
aws-cdk.aws-lambda-event-sources==1.85.0
aws-cdk.aws-s3==1.85.0
aws-cdk.aws-events==1.85.0
aws-cdk.aws-events-targets==1.85.0

# pip install boto3
boto3==1.14.13
botocore==1.17.13
